phidata-agent-serverless-websocket/lambdas/agents/requirements.txt
openai
boto3
phidata